+++ docs/changelog.md
@@ -52,6 +52,7 @@ As features stabilize some brief notes about them will accumulate here.
 * Fixed: mouse reporting gbutton state not cleared when focus is lost, eg: from clicking a link [#744](https://github.com/wez/wezterm/issues/744)
 * Improved: better looking curly underline. Thanks to [@ModProg](https://github.com/ModProg)! [#733](https://github.com/wez/wezterm/pull/733)
 * Fixed: wezterm now sets argv0 to `-$SHELL` to invoke a login shell, rather than running `$SHELL -l`. [#753](https://github.com/wez/wezterm/issues/753)
+* Improved: ssh_config parsing now supports `Match` for `Host`, `LocalUser`.
 
 ### 20210405-110924-a5bb5be8